% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{direction_check}
\alias{direction_check}
\title{Checks if there is a point at given location}
\usage{
direction_check(X, i, edges, vec)
}
\arguments{
\item{X}{A data.frame with two columns, named x and y respectively.}

\item{i}{Numerical. current index.}

\item{edges}{A data.frame with two columns, named x and y respectively.}

\item{vec}{A vector of length 2.}
}
\value{
Logical.
}
\description{
This function checks if one of the rows in X is equal to the sum
of the i-1'th edge and vec.
}
\examples{
edges <- data.frame(x = c(1, 1, 1),
                    y = c(1, 2, 3))
i <- 3
X <- data.frame(x = c(1, 1, 1),
                y = c(1, 2, 3))

vec <- c(0, 1)

direction_check(X, i, edges, vec)
}
