% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtetris.R
\name{ggtetris}
\alias{ggtetris}
\title{Creates a tetris visualization using ggplot2}
\usage{
ggtetris(data, colunms = 10, xlab = NULL, colors = NA, line_color = NA,
  title = NULL)
}
\arguments{
\item{data}{Numerical vector of counts, named or otherwise or data.frame
with a name column and a count column.}

\item{colunms}{Numeric. Number of colunms.}

\item{xlab}{text for below the chart.}

\item{colors}{Same number of colors as values in \code{data}.
If omitted, Color Brewer "Set3" colors are used.}

\item{line_color}{Color of the outline of the polygons in the chart.
Defaults to grey.}

\item{title}{Character. chart title.}
}
\value{
ggplot objects.
}
\description{
This function creates a chart where the number in each group is represented
by a collection of connected squares proportional to the number in that
group. In this function the group counts are supplied as input.
}
\examples{
nums <- c(25, 13, 12, 7, 43)
ggtetris(data = nums)

ggtetris(rep(6, 12), colunms = 8)

ggtetris(c('AA' = 30, 'BB' = 23, 'CC' = 10))

ggtetris(
  data.frame(
    names = LETTERS[1:4],
    vals = c(80, 30, 20, 10)
    ), colunms = 8
 )

 raw_data <- rpois(100, lambda = 1)
 counts <- table(raw_data)
 ggtetris(counts)
}
