% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtetris_tidy.R
\name{ggtetris_tidy}
\alias{ggtetris_tidy}
\title{Tidy ggtetris inputs to regular ggplot usage}
\usage{
ggtetris_tidy(data)
}
\arguments{
\item{data}{Numerical vector of counts, Named or otherwise or data.frame
with a name column and a count column.}
}
\value{
Tidy tibble ready to be used with \code{geom_bar()}.
}
\description{
transforms (named) vector of counts or data.frames with names and counts
to tidy data.frames. If names are not provided for all counts in the vector,
will they be set to capital letters.
}
\examples{
vec <- rep(5, 10)
ggtetris_tidy(vec)

vec_named <- c('AA' = 20, 'BB' = 10, 'CC' = 15)
ggtetris_tidy(vec_named)

data <- data.frame(names = letters[1:4], vals = c(80, 30, 20, 10))
ggtetris_tidy(data)
}
