% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{reduce_matrix}
\alias{reduce_matrix}
\title{Remove redundant points}
\usage{
reduce_matrix(points)
}
\arguments{
\item{points}{A matrix with two columns, named x and y respectively.
Corresponding to the vertices points.}
}
\value{
A matrix with two columns, named x and y respectively.
 Corresponding to the vertices points.
}
\description{
Remove redundant points
}
\examples{
X <- data.frame(x = c(1, 1, 1, 2, 3),
                y = c(1, 2, 3, 2, 2))
points <- tetris(X)
reduce_matrix(points)
}
