% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_tetris.R
\name{scale_tetris}
\alias{scale_colour_tetris}
\alias{scale_color_tetris}
\alias{scale_fill_tetris}
\title{Scales using original Tetris colors as scales}
\usage{
scale_colour_tetris(..., palette = "Vadim Gerasimov")

scale_color_tetris(..., palette = "Vadim Gerasimov")

scale_fill_tetris(..., palette = "Vadim Gerasimov")
}
\arguments{
\item{...}{Other arguments passed on to [discrete_scale()] to control name, limits,
breaks, labels and so forth.}

\item{palette}{Character. Name of color scale. Pick between "Vadim Gerasimov",
"Microsoft", "Sega", "The New Tetris", "The Tetris Company", "Atari",
"The Soviet Mind Game", "Tetris DX", "Super Tetris 3", "Tetris Plus"
and "The Next Tetris". Defaults to "Vadim Gerasimov".}
}
\description{
Allows to set scales to the colors used in an array of different Tetris
games. See \url{https://en.wikipedia.org/wiki/Tetris#Tetromino_colors}
for more information of names and color. Name of color scale are
"Vadim Gerasimov", "Microsoft", "Sega", "The New Tetris",
"The Tetris Company", "Atari", "The Soviet Mind Game", "Tetris DX",
"Super Tetris 3", "Tetris Plus" and "The Next Tetris". Defaults to
"Vadim Gerasimov". Will pad with white if after 7 colors are used.
}
\examples{
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
(d <- ggplot(dsamp, aes(carat, price)) +
  geom_point(aes(colour = color)))
  d + scale_colour_tetris()
}
