% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{square_it_mat}
\alias{square_it_mat}
\title{Expands unit-squares on points.}
\usage{
square_it_mat(X)
}
\arguments{
\item{X}{A data.frame with two columns, named x and y respectively.}
}
\value{
Takes a set of gridpoints and returns the the vertices of the unit
    square it the points as its center.
}
\description{
Expands unit-squares on points.
}
\examples{
X <- data.frame(x = c(1, 1, 1),
                y = c(1, 2, 3))
square_it_mat(X)
}
