% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetris.R
\name{tetris}
\alias{tetris}
\title{Calculate the tetris shape outlining a set of connected points.}
\usage{
tetris(points)
}
\arguments{
\item{points}{A data.frame or matrix with two columns, named x and y
respectively.}
}
\value{
A matrix with two columns, named x and y respectively. Corresponding
 to the vertices points.
}
\description{
Calculate the tetris shape outlining a set of connected points.
}
\examples{
X <- data.frame(x = c(1, 1, 1, 2, 3),
                y = c(1, 2, 3, 2, 2))
tetris(X)
}
