% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetris_colors.R
\name{tetris_color}
\alias{tetris_color}
\title{Tetris colors}
\usage{
tetris_color(palette = "Vadim Gerasimov", n = 7)
}
\arguments{
\item{palette}{Character. Name of color scale. Pick between "Vadim Gerasimov",
"Microsoft", "Sega", "The New Tetris", "The Tetris Company", "Atari",
"The Soviet Mind Game", "Tetris DX", "Super Tetris 3", "Tetris Plus"
and "The Next Tetris". Defaults to "Vadim Gerasimov".}

\item{n}{Integer. Number of colors. Maximum of 7 available. Defaults to 7.}
}
\value{
Numerical.
}
\description{
Returns a vector of colors that match the colors of the Tetromino pieces
in a range of different Tetris editions. See
\url{https://en.wikipedia.org/wiki/Tetris#Tetromino_colors}
for more information of names and color. Name of color palettes are
"Vadim Gerasimov", "Microsoft", "Sega", "The New Tetris",
"The Tetris Company", "Atari", "The Soviet Mind Game", "Tetris DX",
"Super Tetris 3", "Tetris Plus" and "The Next Tetris". Defaults to
"Vadim Gerasimov". Will pad with white if after 7 colors are used.
}
\examples{
tetris_color(7)
tetris_color("Microsoft", n = 4)
}
